iommu: qcom: iommu-v1/sec: Fix race unclocked access for secure IOMMU
There is a race condition that makes the secure IOMMU driver
to call an IOMMU access while clocks are possibly down:
the insecure driver has got the same clocks as the secure part
and, when the first one does an operation that requires clocks
to be enabled, it will disable it after finishing that.
The secure IOMMU never tries to manage clocks, hence, everytime
the insecure IOMMU shuts down clocks, it will be very likely
for the secure driver to produce an unclocked access.

To fix this race condition, add two functions to the IOMMU
insecure driver where clocks are managed mainly, then call
those to manage clocks in the secure driver using proper
locking.