arm64: DT: MSM8956: Add support for efuse based GFX clock
Add support to read the efuse register and determine the version
and based on the version load the corresponding frequency-voltage
mapping table.
Also update the fmax & list of supported frequencies based on the
version.

This was implemented in clock-gcc-8976 driver, but fmax & list of
supported frequencies was missed in dts.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>